[update] DnD機能と検索機能のstateのリフトアップ #40
